requirements.txt
numpy==1.26.2
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
Pillow==10.0.1
Pillow==10.1.0
pytube==15.0.0
streamlit==1.27.2
tensorflow==2.10.0
